Changed int to uint

+++ pkg/simulation/simulation.go
@@ -39,14 +39,14 @@ func (s *Simulation) ResetTrams() {
 	}
 }
 
-func (s *Simulation) FetchData(url string, tramWorkerCount int) {
+func (s *Simulation) FetchData(url string, tramWorkerCount uint) {
 	s.city.FetchCityData(url)
 	s.controlCenter = controlcenter.NewControlCenter(s.city)
 	s.ResetTrams()
 	s.tramWorkersData.reset(len(s.trams))
 
 	if tramWorkerCount == 0 {
-		tramWorkerCount = runtime.NumCPU() * 11 / 10
+		tramWorkerCount = uint(runtime.NumCPU()) * 11 / 10
 	}
 
 	for range tramWorkerCount {
